move patch to legacy and ASF s3 provider
